refactored code to manage json formatting in one place, remove classes where we aren't encapsulating data, structure package contents in a more meaningful way, and stub out the configuration data for now so we can access the database.
