+++ webserver/src/main/manager/UnitOfMeasure.py
@@ -22,37 +22,11 @@
 #   along with this program.  If not, see <http://www.gnu.org/licenses/>.      #
 ################################################################################
 
-from dao import DatabaseHandler
-from manager import dateTimeUtility
-
-class PortManager ():
-
-    #TODO:
-    # Port Manager: calls necessary managers and utilities to generate parameters for sql.
-    #  Mgr for reading config files to know what table /column names there are.
-    #  Utility for calc time range
-    #
-    #
-    validPortNumbers = ()
-
-    def __init__(self):
-        # TODO:  get from config file
-        self.validPortNumbers = (8023, 8082, 8083)
+from enum import Enum
 
 
-    def isPortValid(self, port_number):
-        if (port_number in self.validPortNumbers):
-            return True
-        else:
-            return False
-
-    def getPort(self, port_number, uom, unit):
-        print("Retrieving port:" + str(port_number))
-
-        if self.isPortValid(port_number):
-            # return self.fakePort80DataWithTime
-            return DatabaseHandler.getJson(port_number, uom, unit )
-        else:
-            return None
-
-
+class UnitOfMeasure(Enum):
+    MINUTE = "minutes"
+    HOUR = "hours"
+    DAY = "days"
+    WEEK = "weeks"
