+++ webserver/src/main/manager/utilities.py
@@ -22,41 +22,18 @@
 #   along with this program.  If not, see <http://www.gnu.org/licenses/>.      #
 ################################################################################
 
-'''
-This class will handle any date/time requests from the webserver.
-This includes calculating the date/time range for data to return,
-as well as converting the string timestamp in the database.
+from datetime import datetime, timedelta
 
-Author: Charlie Mitchell
-Last Revised: 28 February, 2016
-'''
-
-import datetime
-from server.UnitOfMeasure import UnitOfMeasure
+from manager.UnitOfMeasure import UnitOfMeasure
 
 
-class DateTimeManager:
-    # Return the date for how far back to query DB.
+def getIntValue(givenStr):
+    print("given str is: " + givenStr)
+    try:
+        returnval = int(givenStr)
+        return returnval
+    except Exception as e:
+        print("received invalid string to convert to int: " + givenStr)
+        print (str(e))
+        return None
 
-    def get_begin_date(self, unit, unit_size):
-
-        if unit == UnitOfMeasure.MINUTE:
-            d = datetime.timedelta(minutes=unit_size)
-        elif unit == UnitOfMeasure.HOUR:
-            d = datetime.timedelta(hours=unit_size)
-        elif unit == UnitOfMeasure.DAY:
-            d = datetime.timedelta(days=unit_size)
-        elif unit == UnitOfMeasure.WEEK:
-            d = datetime.timedelta(weeks=unit_size)
-        else:
-            d=1;  #defaults to 1 day but should never happen
-
-        return self.calc_date(d)
-
-    def calc_date(self, delta):
-        now = datetime.datetime.now()
-        return now - delta
-
-    # Takes the datetime object and returns a string in ISO 8601 format.
-    def get_iso_format(self, begin_date):
-        return begin_date.isoformat()
