--- webserver/src/main/runServer.sh
@@ -1,3 +1,0 @@
-export PYTHONPATH=${PYTHONPATH}:$PWD
-echo $PYTHONPATH
-python3 ./server/main.py
