+++ webserver/src/main/server/PortsServiceHandler.py
@@ -1,8 +1,7 @@
 from manager.PortManager import PortManager
 from server.RESTRequestHandler import RestRequestHandler
-from server.UnitOfMeasure import UnitOfMeasure
 
-# Todo:  get json from database data
+
 # Todo:  set up method for each path
 # Todo:  decide what we want for path names (what the api will be)
 
@@ -10,7 +9,7 @@ from server.UnitOfMeasure import UnitOfMeasure
 class PortsServiceHandler (RestRequestHandler):
 
     def getPortData(self, portnumber):
-        self.getPortDataByTime(portnumber, UnitOfMeasure.DAY, 1)
+        self.getPortDataByTime(portnumber, "days", 1)
 
     def getPortDataByTime(self, portnumber, uom, unit):
         portmgr = PortManager()
