--- webserver/src/main/server/UnitOfMeasure.py
@@ -1,8 +1,0 @@
-from enum import Enum
-
-
-class UnitOfMeasure(Enum):
-    MINUTE = "minutes"
-    HOUR = "hours"
-    DAY = "days"
-    WEEK = "weeks"
